Refactor Dialog to use Popover
