+++ front/components/assistant/conversation/content/ClientExecutableRenderer.tsx
@@ -2,7 +2,6 @@ import {
   Button,
   CodeBlock,
   CommandLineIcon,
-  LinkIcon,
   SparklesIcon,
   Spinner,
 } from "@dust-tt/sparkle";
@@ -14,7 +13,7 @@ import { isFileUsingConversationFiles } from "@app/lib/files";
 import { useFileContent } from "@app/lib/swr/files";
 import type { ConversationType, LightWorkspaceType } from "@app/types";
 
-import { ShareInteractiveFileDialog } from "../ShareInteractiveFileDialog";
+import { ShareInteractiveFilePopover } from "../ShareInteractiveFilePopover";
 import { useInteractiveContentContext } from "./InteractiveContentContext";
 import { InteractiveContentHeader } from "./InteractiveContentHeader";
 
@@ -42,7 +41,6 @@ export function ClientExecutableRenderer({
   );
 
   const [showCode, setShowCode] = React.useState(false);
-  const [isShareDialogOpen, setIsShareDialogOpen] = React.useState(false);
 
   if (isFileContentLoading) {
     return (
@@ -62,66 +60,56 @@ export function ClientExecutableRenderer({
   }
 
   return (
-    <>
-      <ShareInteractiveFileDialog
-        fileId={fileId}
-        isOpen={isShareDialogOpen}
-        onClose={() => setIsShareDialogOpen(false)}
-        owner={owner}
-      />
-      <div className="flex h-full flex-col">
-        <InteractiveContentHeader
-          title={fileName || "Client Executable"}
-          subtitle={fileId}
-          onClose={closeContent}
-        >
-          <Button
-            icon={showCode ? SparklesIcon : CommandLineIcon}
-            onClick={() => setShowCode(!showCode)}
-            size="xs"
-            tooltip={showCode ? "Switch to Rendering" : "Switch to Code"}
-            variant="outline"
-          />
-          <Button
-            icon={LinkIcon}
-            onClick={() => setIsShareDialogOpen(true)}
-            size="xs"
-            variant="ghost"
-            tooltip={
-              isUsingConversationFiles
-                ? "This interactive file uses conversation files. It cannot be shared publicly."
-                : "Share public link"
-            }
-            disabled={isUsingConversationFiles}
-          />
-        </InteractiveContentHeader>
+    <div className="flex h-full flex-col">
+      <InteractiveContentHeader
+        title={fileName || "Client Executable"}
+        subtitle={fileId}
+        onClose={closeContent}
+      >
+        <Button
+          icon={showCode ? SparklesIcon : CommandLineIcon}
+          onClick={() => setShowCode(!showCode)}
+          size="xs"
+          tooltip={showCode ? "Switch to Rendering" : "Switch to Code"}
+          variant="outline"
+        />
+        <ShareInteractiveFilePopover
+          fileId={fileId}
+          owner={owner}
+          disabled={isUsingConversationFiles}
+          tooltip={
+            isUsingConversationFiles
+              ? "This interactive file uses conversation files. It cannot be shared publicly."
+              : "Share public link"
+          }
+        />
+      </InteractiveContentHeader>
 
-        {/* Content */}
-        <div className="flex-1 overflow-hidden">
-          {showCode ? (
-            <div className="h-full overflow-auto px-4">
-              <CodeBlock wrapLongLines className="language-tsx">
-                {fileContent}
-              </CodeBlock>
-            </div>
-          ) : (
-            <div className="h-full">
-              <VisualizationActionIframe
-                agentConfigurationId={null}
-                workspaceId={owner.sId}
-                visualization={{
-                  code: fileContent ?? "",
-                  complete: true,
-                  identifier: `viz-${fileId}`,
-                }}
-                key={`viz-${fileId}`}
-                conversationId={conversation.sId}
-                isInDrawer={true}
-              />
-            </div>
-          )}
-        </div>
+      {/* Content */}
+      <div className="flex-1 overflow-hidden">
+        {showCode ? (
+          <div className="h-full overflow-auto px-4">
+            <CodeBlock wrapLongLines className="language-tsx">
+              {fileContent}
+            </CodeBlock>
+          </div>
+        ) : (
+          <div className="h-full">
+            <VisualizationActionIframe
+              agentConfigurationId={null}
+              workspaceId={owner.sId}
+              visualization={{
+                code: fileContent ?? "",
+                complete: true,
+                identifier: `viz-${fileId}`,
+              }}
+              key={`viz-${fileId}`}
+              conversationId={conversation.sId}
+              isInDrawer={true}
+            />
+          </div>
+        )}
       </div>
-    </>
+    </div>
   );
 }
